Rename example configuration portlet, set required=false in configuration interface's attribute definitions
