+++ develop/tutorials/code/liferay-plugins-sdk-7.0.0/portlets/example-config-portlet/ivy.xml
